cleanup: fixed styling + renamed props
