+++ clients/search-component/src/TrieveModal/index.tsx
@@ -15,7 +15,7 @@ import r2wc from "@r2wc/react-to-web-component";
 import { setClickTriggers } from "../utils/hooks/setClickTriggers";
 import { ChunkGroup } from "trieve-ts-sdk";
 import { FloatingActionButton } from "./FloatingActionButton";
-import { FloatingIconButton } from "./FloatingIconButton";
+import { FloatingSearchIcon } from "./FloatingSearchIcon";
 
 const Modal = () => {
   useKeyboardNavigation();
@@ -172,7 +172,7 @@ const Modal = () => {
           </div>
         </>
       )}
-      <FloatingIconButton />
+      <FloatingSearchIcon />
       <FloatingActionButton />
     </>
   );
